+++ CHANGELOG.md
@@ -108,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed other macros and definitions used internally.
 - Changed type of the first parameter of `lt_random_bytes` to `lt_handle_t`.
 - Replaced various numeric constants with appropriate macro constants and sizeof operators.
+- `lt_get_info_fw_bank()`: Renamed parameter `max_len` to `max_size`.
 
 ### Added
 - CMake option for setting logging verbosity level: `LT_LOG_LVL`.
@@ -118,6 +119,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Constants for length of P256 and ED25519 private and public keys.
 - Constant for length of ECDSA/EDDSA signature.
 - Constants for lengths of L2 request fields (ID, LEN, CRC).
+- `lt_get_info_fw_bank()`: Parameter `read_size` to indicate the number of read bytes from the FW bank.
 
 ### Fixed
 - `lt_r_mem_data_write()`, `lt_out__r_mem_data_write()`: Mark `data` as `const`.